Add configurable optimization for versioned batched updates.
When Solr versioning is enabled the default traject json writer often
has to break up the batched update into many single updates do to Solr
responding with 409 errors.

This commit adds a configurable optimization option which when enabled
will make a call to Solr to find the record update dates of records that
are about to be batched processed. Then it uses this information to
filter out the batched records if they are older than the Solr records.
Thus, we avoid Solr responding with 409 and avoid breaking up the
batched update into many single updates.

The reason this is configurable is because we wouldn't need or want to
do this for full re-indexing jobs which would happen on a clean Solr
collection.

This optimization does not guarantee to work when multiple workers are
indexing simultaneously.  But, it's still better than nothing.