Add hathi_bib_key_display for relvant items
Adds a csv file in `lib/hathi_data_overlap.csv` that contains oclc
number and Hathi Trust bib_key for physical items in the Temple
collection that have been digitized by Hathi and are available under the
temporary emergency access program.

This change causes traject to search (grep) the overlap.csv file for the
records oclc number. If it returns nothing, no record is created. If it
matches, it returns a string with "bib_key,oclc_num".

This will slow down the indexing process as grepping a 12MB file takes
~.35 seconds, which is substantally slower than a usual traject process.