Delete suppressed documents from the solr catatlog
When not in FULL_REINDEX context we do not skip items for ingest,
instead we flip a suppression field on the Solr document that is about
to be indexed and then we rely on filtering that document using Solr
configuration at query time.

We do this because merely skipping index would not account for the
documents already present in the Solr database and we have always
assumed that the task of deleting the suppressed items would be unduly
complicated.

Since we have implemented our own delete feature for cob_index (for
another issue) we can easily leverage that new feature to simply delete
documents that would otherwise be suppressed.