Refactor flank macro.
The flank macro is used to add a token to the beginning  and end of
a string that is going to be indexed. But, with the version this change
fixes, that token is always the same and for Solr tokenizers that split on
white spaces it's essentially lost in the shuffle (just another token).

By updating the macro to create a token that embeds more information
about what exactly a string begins with and a string endswith we can
include the specific token to pull the desired results.
``